# Conflicts:
#	src/main/java/de/unipassau/sep19/hafenkran/clusterservice/controller/ExperimentController.java
#	src/main/java/de/unipassau/sep19/hafenkran/clusterservice/repository/ExperimentRepository.java
#	src/main/java/de/unipassau/sep19/hafenkran/clusterservice/service/ExperimentService.java
#	src/main/java/de/unipassau/sep19/hafenkran/clusterservice/service/impl/ExperimentServiceImpl.java
#	src/test/de/unipassau/sep19/hafenkran/clusterservice/service/impl/ExperimentServiceImplTest.java

+++ src/main/java/de/unipassau/sep19/hafenkran/clusterservice/service/impl/ExecutionServiceImpl.java
@@ -1,6 +1,11 @@
 package de.unipassau.sep19.hafenkran.clusterservice.service.impl;
 
-import de.unipassau.sep19.hafenkran.clusterservice.dto.*;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionCreateDTO;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionDTO;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionDTOList;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.ResultDTO;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.StdinDTO;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.UserDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.exception.ResourceNotFoundException;
 import de.unipassau.sep19.hafenkran.clusterservice.kubernetesclient.KubernetesClient;
 import de.unipassau.sep19.hafenkran.clusterservice.model.ExecutionDetails;
@@ -23,7 +28,11 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 import java.util.regex.Pattern;
 
 @Slf4j
@@ -90,8 +99,11 @@ public class ExecutionServiceImpl implements ExecutionService {
      */
     @Override
     public ExecutionDetails createExecution(@NonNull ExecutionDetails executionDetails) {
+        executionDetails.validatePermissions();
+
         final ExecutionDetails savedExecutionDetails =
                 executionRepository.save(executionDetails);
+
 
         log.info(String.format("Execution with id %s created",
                 savedExecutionDetails.getId()));
@@ -106,6 +118,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     public ExecutionDTO terminateExecution(@NonNull UUID executionId) {
 
         ExecutionDetails executionDetails = getExecutionDetails(executionId);
+
+        executionDetails.validatePermissions();
 
         try {
             kubernetesClient.deletePod(executionDetails);
@@ -148,7 +162,10 @@ public class ExecutionServiceImpl implements ExecutionService {
 
     private ExecutionDetails retrieveExecutionDetailsById(@NonNull UUID id) {
         final Optional<ExecutionDetails> execution = executionRepository.findById(id);
-        return execution.orElseThrow(() -> new ResourceNotFoundException(ExecutionDetails.class, "id", id.toString()));
+        ExecutionDetails executionDetails = execution.orElseThrow(
+                () -> new ResourceNotFoundException(ExecutionDetails.class, "id", id.toString()));
+        executionDetails.validatePermissions();
+        return execution.get();
     }
 
     /**
@@ -232,6 +249,8 @@ public class ExecutionServiceImpl implements ExecutionService {
                 executionRepository.findById(executionId).orElseThrow(
                         () -> new ResourceNotFoundException(ExecutionDetails.class, "id", executionId.toString()));
 
+        executionDetails.validatePermissions();
+
         if (status.equals(Status.FINISHED)) {
             executionDetails.setTerminatedAt(LocalDateTime.now());
         }
@@ -246,6 +265,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     }
 
     private ExecutionDetails startExecution(@NonNull ExecutionDetails executionDetails) {
+        executionDetails.validatePermissions();
+
         final String podName;
         try {
             podName = kubernetesClient.createPod(executionDetails);
@@ -348,8 +369,10 @@ public class ExecutionServiceImpl implements ExecutionService {
     public ExecutionDetails getExecutionOfPod(@NonNull String podName, @NonNull UUID namespace) {
 
         ExperimentDetails experiment = experimentRepository.findById(namespace).orElseThrow(
-                () -> new ResourceNotFoundException(ExperimentDetails.class, "id",
-                        namespace.toString()));
+                () -> new ResourceNotFoundException(ExperimentDetails.class, "id",namespace.toString()));
+
+        experiment.validatePermissions();
+
         return executionRepository.findByPodNameAndExperimentDetails(podName, experiment);
     }
 
@@ -367,6 +390,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     public ExecutionDTO deleteExecution(@NonNull UUID executionId) {
 
         ExecutionDetails executionDetails = getExecutionDetails(executionId);
+
+        executionDetails.validatePermissions();
 
         if (executionDetails.getStatus().equals(Status.RUNNING)) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Can not delete executions in running");
